update edited version
