+++ static/layout.css
@@ -1,0 +1,87 @@
+*{
+    margin: 0px;
+}
+body{
+    top: 0px;
+    left: 0px;
+}
+.topLayout{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+.top {
+    position: relative;
+    display: flex;
+    width: 1200px;
+    height: 54px;
+}
+.topleft{
+    font-size: 30px;
+    line-height: 54px;
+}
+.tpTrip{
+    font-size: 30px;
+    color: #448899;
+    line-height: 54px;
+}
+.topRight{
+    position: absolute;
+    display: flex;
+    font-size: 16px;
+    align-items: center;
+    right: 0px;
+    line-height: 54px;
+}
+.login{
+    margin-left: 5px;
+}
+.downLayout{
+    background-color: #757575;
+    color: #f8f8f8;
+    text-align: center;
+    line-height: 104px;
+    height: 104px;
+}
+.welcome{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+.welcomeImg{
+    width: 100%;
+    height: 320px;
+    object-fit: cover;
+    object-position: top;
+}
+.searchBox{
+    width: 1200px;
+    position: absolute;
+}
+.searchTitle{
+    color: #f8f8f8;
+    font-size: 28px;
+}
+.searchComment{
+    color: #f8f8f8;
+    font-size: 16px;
+}
+.searchHere{
+    display: flex;
+}
+.searchSearch{
+    height: 46px;
+    width: 400px;
+    border-width: 0px;
+    padding: 0px;
+}
+.input,input::placeholder{
+    padding-left: 15px;
+}
+.searchButton{
+    height: 46px;
+    width: 60px;
+    background: #448899;
+    border-width: 0px;
+    cursor: pointer;;
+}
