+++ templates/index.html
@@ -1,0 +1,31 @@
+{% extends "layout.html" %}
+
+{% block title %}
+    台北一日遊
+{% endblock %}
+
+{% block main %}
+
+    <div>
+        <div class="welcome">
+            <img class="welcomeImg" src="/static/welcome.png" alt=""/>                    
+            <div class="searchBox">
+                <div class="searchTitle">輕鬆享受台北一日悠閒</div>
+                <br/>
+                <div class="searchComment">探索每個角落，體驗城市的深度旅遊行程</div>
+                <br/>
+                <div class="searchHere">
+                    <input class="searchSearch" name="search" placeholder="請輸入景點名稱查詢" type="text"/>
+                    <button class="searchButton" type="submit">
+                        <img src="/static/icon_search.png" alt=""/>
+                    </button>
+                </div>    
+            </div>
+        </div>
+        <div class="attractionsBox">
+            
+        </div>
+    </div>
+
+
+{% endblock %}
