+++ templates/layout.html
@@ -1,0 +1,37 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <link rel="stylesheet" type="text/css" href="/static/layout.css" />
+    <!-- CSS only -->
+    <script src="https://unpkg.com/react@18/umd/react.development.js" crossorigin></script>
+    <script src="https://unpkg.com/react-dom@18/umd/react-dom.development.js" crossorigin></script>
+    <script src="https://unpkg.com/babel-standalone@6/babel.min.js"></script>
+    <title>{% block title %}{% endblock %}</title>
+</head>
+<body>
+    <div class="topLayout">
+        <div class="top">
+            <div class="topLeft">
+                <div class="tpTrip">台北一日遊</div>
+            </div>
+            <div class="topRight">
+                <div class="project">預定行程</div>
+                <div class="login">登入/註冊</div>    
+            </div>
+        </div>
+    </div>
+    <!-- {% if session.get("userID") %}
+
+    {% else %}
+
+    {% endif %}-->
+    <main>
+        {% block main %}{% endblock %}
+    </main> 
+    <div class="downLayout">
+        <div class="copyright">COPYRIGHT © 2021 台北一日遊</div>
+    </div>
+</body>
