--- test_hello.py
@@ -1,1 +1,0 @@
-print("hello")
